+++ src/Reporting/Formatter/Ansi/AnsiColor.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Qualimetrix\Reporting\Formatter\Support;
+namespace Qualimetrix\Reporting\Formatter\Ansi;
 
 /**
  * Lightweight ANSI color wrapper for formatter output.